Move editor tabs actions from src/actions/sources to src/actions
