+++ src/actions/sources/index.js
@@ -8,4 +8,3 @@ export * from "./loadSourceText";
 export * from "./newSources";
 export * from "./prettyPrint";
 export * from "./select";
-export * from "./tabs";
